// packages/mdoc/mdoc-client/src/mdoc/model/presentation-definition.ts
export interface PresentationDefinitionField {
  path: string[];
  intent_to_retain: boolean;
}

export interface Format {
  mso_mdoc: {
    alg: string[];
  };
}

export interface InputDescriptor {
  id: string;
  format: Format;
  constraints: {
    limit_disclosure: string;
    fields: PresentationDefinitionField[];
  };
}

export interface PresentationDefinition {
  id: string;
  input_descriptors: InputDescriptor[];
}